allow iteration of url variable in consume tag
Allow iteration of the url referenced by the consume tag. To allow this
we have pull the url out of the context each time we render. Also to
allow the page fragment cache to work correctly our cache key now has
to be a combination of the target and url. Currently the path taken with
the timeout won't quite work as expected I believe. However, I don't quite
understand the actual life cycle of the tags object hash to make the
change at this stage.