Fix entity tags (ETags)
ETag header should always be set, otherwise Rails does it based on the content rather than our preferred object.

ETag header should not be set based on the Page object alone, as the page object does not account for differences in output between locales. This means that multiple paths based on locales can have the same ETag regardless of path and/or content. This change adds selected params into the object being passed to `fresh_when` to create the ETag hash.