Updated inline editor layout
Changed the inline editor layout such that #page is guaranteed to take up the entire viewport. This solves a problem where a fixed-position HTML root tag in the iframe's document would cause the edit page to size incorrectly. Forcing the #page element and therefore the iframe inside it to take up the full viewport seems a sensible way to avoid this, and should not have any adverse effects.